+++ tests/test_printer.py
@@ -77,6 +77,7 @@ class TestPrinter(unittest.TestCase):
             "a.foo()",
             "a:foo().bar()",
             "f=function(x,...)end",
+            "x={['foo/bar']=1}"
         ]
         for a in cases:
             with self.subTest(code=a):